[BE] refactor: Entity 필드 수정
- Entity에 createdBy, createdAt, modifiedBy, modifiedAt 필드 추가

#83
