+++ BE/src/main/java/kr/codesquad/issuetracker07/controller/MilestoneController.java
@@ -1,7 +1,7 @@
 package kr.codesquad.issuetracker07.controller;
 
-import kr.codesquad.issuetracker07.domain.Issue;
-import kr.codesquad.issuetracker07.domain.Milestone;
+import kr.codesquad.issuetracker07.entity.Issue;
+import kr.codesquad.issuetracker07.entity.Milestone;
 import kr.codesquad.issuetracker07.dto.MilestoneVO;
 import kr.codesquad.issuetracker07.response.MilestoneResponse;
 import kr.codesquad.issuetracker07.service.IssueService;
