--- BE/src/main/java/kr/codesquad/issuetracker07/domain/Assignee.java
@@ -1,26 +1,0 @@
-package kr.codesquad.issuetracker07.domain;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Assignee {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String imageUrl;
-
-    @ManyToOne
-    @JoinColumn(name = "issue_id")
-    private Issue issue;
-}
