+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/Assignee.java
@@ -1,28 +1,28 @@
-package kr.codesquad.issuetracker07.domain;
+package kr.codesquad.issuetracker07.entity;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
 @Entity
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Attachment {
+public class Assignee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private boolean isAttached;
+    private String name;
+
+    private String imageUrl;
 
     @ManyToOne
     @JoinColumn(name = "issue_id")
     private Issue issue;
-
-    @ManyToOne
-    @JoinColumn(name = "label_id")
-    private Label label;
 }
