+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/Attachment.java
@@ -1,0 +1,30 @@
+package kr.codesquad.issuetracker07.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Attachment {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private boolean isAttached;
+
+    @ManyToOne
+    @JoinColumn(name = "issue_id")
+    private Issue issue;
+
+    @ManyToOne
+    @JoinColumn(name = "label_id")
+    private Label label;
+}
