+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/AuthProvider.java
@@ -1,4 +1,4 @@
-package kr.codesquad.issuetracker07.domain;
+package kr.codesquad.issuetracker07.entity;
 
 public enum AuthProvider {
     local,
