+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/Comment.java
@@ -1,15 +1,16 @@
-package kr.codesquad.issuetracker07.domain;
+package kr.codesquad.issuetracker07.entity;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
