+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/Issue.java
@@ -1,16 +1,17 @@
-package kr.codesquad.issuetracker07.domain;
+package kr.codesquad.issuetracker07.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
@@ -26,7 +27,11 @@ public class Issue {
 
     private boolean isOpen;
 
+    private String createdBy;
+
     private LocalDateTime createdAt;
+
+    private String modifiedBy;
 
     private LocalDateTime modifiedAt;
