+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/Label.java
@@ -1,13 +1,16 @@
-package kr.codesquad.issuetracker07.domain;
+package kr.codesquad.issuetracker07.entity;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
@@ -22,6 +25,14 @@ public class Label {
     private String description;
 
     private String backgroundColor;
+
+    private String createdBy;
+
+    private LocalDateTime createdAt;
+
+    private String modifiedBy;
+
+    private LocalDateTime modifiedAt;
 
     @OneToMany(mappedBy = "label")
     List<Attachment> attachmentList;
