+++ BE/src/main/java/kr/codesquad/issuetracker07/entity/User.java
@@ -1,4 +1,4 @@
-package kr.codesquad.issuetracker07.domain;
+package kr.codesquad.issuetracker07.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
@@ -6,14 +6,9 @@ import lombok.*;
 import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.Id;
 
 @Entity
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
