+++ BE/src/main/java/kr/codesquad/issuetracker07/repository/AddingRepository.java
@@ -1,6 +1,6 @@
 package kr.codesquad.issuetracker07.repository;
 
-import kr.codesquad.issuetracker07.domain.Adding;
+import kr.codesquad.issuetracker07.entity.Adding;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
